+++ src/tightdb/table_accessors.hpp
@@ -1280,7 +1280,7 @@ public:
         return *Base::m_query;
     };
 
-    float sum(std::size_t* resultcount=NULL, std::size_t start=0,
+    double sum(std::size_t* resultcount=NULL, std::size_t start=0,
                std::size_t end = std::size_t(-1), std::size_t limit=std::size_t(-1)) const
     {
         return Base::m_query->m_impl.sum_float(col_idx, resultcount, start, end, limit);
